+++ Units/parser-python.r/strings.d/input.py
@@ -1,0 +1,37 @@
+
+VAR01 = ""
+VAR02 = "hello"
+VAR03 = ''
+VAR04 = 'hi'
+VAR05 = "\""
+VAR06 = '\''
+VAR07 = "'''"
+VAR08 = '"""'
+
+VAR21 = r""
+VAR22 = u"hello"
+VAR23 = R''
+VAR24 = U'hi'
+VAR25 = b"\""
+VAR26 = B'\''
+VAR27 = br"'''"
+VAR28 = Br'"""'
+VAR29 = bR"hello"
+VAR30 = BR"hello"
+# Python >= 3.3
+VAR31 = rb"hello"
+VAR32 = rB"hello"
+VAR33 = Rb"hello"
+VAR34 = RB"hello"
+# Python 2 only
+VAR35 = ur"hello"
+VAR36 = UR"hello"
+VAR37 = uR"hello"
+VAR38 = Ur"hello"
+
+# also test not-too-bad behavior when strings are not terminated properly
+VAR80 = "oops
+VAR81 = 'oopsie
+VAR82 = "\"
+VAR83 = '\'
+VAR84 = 42