+++ Units/parser-python.r/triple-quotes.d/input.py
@@ -1,0 +1,12 @@
+
+VAR1 = '''
+def bug1():
+    pass
+''
+def bug2():
+    pass
+'''
+
+VAR2 = """ "" ""\" "\"" \"\"\"\\"""
+
+ZZZ = True