+++ docs/news.rst
@@ -27,6 +27,7 @@ Fully rewritten parsers
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * C (see :ref:`The new C/C++ parser <cxx>`)
 * C++ (see :ref:`The new C/C++ parser <cxx>`)
+* Python (see :ref:`The new Python parser <python>`)
 
 New parsers
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~